Refactor repository static page modules
