+++ docs/agent-audit/reasoning/2026/06/08/repository-review-workflows-trace.md
@@ -50,6 +50,10 @@
 - `src/govuk/templates/pages/repository.njk`
 - `src/styles/repository.scss`
 - `public/js/repository-static-page.js`
+- `public/js/repository-static/browse.js`
+- `public/js/repository-static/candidate.js`
+- `public/js/repository-static/review.js`
+- `public/js/repository-static/shared.js`
 - `infra/cloudflare/src/service/repository.js`
 - `infra/cloudflare/src/service/index.js`
 - `infra/cloudflare/src/worker.js`
@@ -80,6 +84,8 @@
 - Limited review queue responses to 10 items per page and added queue pagination support through the review API and page controller.
 - Reworked the tab implementation so each GOV.UK tab panel contains its own queue-specific workbench shell, with unique list, detail and pagination targets.
 - Wired review tab selection to update the page title, lead paragraph, supporting body text, breadcrumb current item and document title so the whole page context follows the selected queue.
+- Refactored the 946-line `public/js/repository-static-page.js` into a small module entrypoint plus shared, browse, candidate and review modules.
+- Checked adjacent repository frontend scripts and kept `repository-page.js` and `repository-artefact-page.js` unchanged because they are smaller and not directly part of the expanding static review/browse implementation.
 - Removed static review-check, review-outcome and withdrawal-reason copy from the review route page data.
 - Aligned queue and selected-record panels with matching bordered containers so their top borders land on the same line.
 - Added queue and action tests at both route-state and runtime levels.
@@ -99,6 +105,7 @@
 - `npm run lint`
 - `npm run format:check`
 - Local browser check for Candidate, Due review and Withdrawn review tab page-context switching
+- Local browser check after the static repository module split
 
 ## Validation results
 
@@ -110,6 +117,7 @@
 - Format check: passed
 - Lint: passed with existing repository warnings only
 - Local browser check confirmed the H1, lead, body text, breadcrumb current item and document title update when switching from Due review to Candidate artefacts and Withdrawn artefacts.
+- Local browser check confirmed the refactored module entrypoint still loads and Candidate tab switching still updates page context and visible panel state.
 
 ## Issues and pivots
 
@@ -121,6 +129,7 @@
 - The first tabs pass still left inactive queues as placeholder panels, which is why only one tab displayed real workbench content. The tab panels were then expanded to provide queue-specific workbench shells inside each GOV.UK tab panel.
 - Moving a single workbench node between tab panels proved brittle with the GOV.UK tabs runtime. The implementation now renders one shell per queue and scopes script updates to `[data-review-workbench]`, avoiding both empty selected panels and duplicate IDs.
 - A later review found that the workbench tab changed but the page-level route context remained on the initially loaded queue. The controller now updates the route-level copy and document metadata as part of the active-tab state.
+- The static repository script had grown to 946 lines while accumulating browse, candidate and review behaviours. It was split into focused modules before further review work was added.
 
 ## Residual risks
